🎉 v2.2.1 | Redbox Locator
This release features a quick fix with the maxZoom for clustering. It allows for an easier view on mobile devices.

Warning: when accessing the map on iOS, it's highly recommended that you avoid Safari, because issues with memory overflow arise. Apple has yet to fix this issue, so be cautious!